refactor(agent)!: rename GroundingDecodable to GroundingRepresentable
