+++ Sources/SwiftAgent/Agent/GroundingRepresentable.swift
@@ -1,0 +1,6 @@
+// By Dennis Müller
+
+import Foundation
+
+public protocol GroundingRepresentable: Sendable, Equatable, Codable {}
+
