--- Sources/SwiftAgent/Prompt Context/PromptContext.swift
@@ -1,6 +1,0 @@
-// By Dennis Müller
-
-import Foundation
-
-public protocol GroundingDecodable: Sendable, Equatable, Codable {}
-
