+++ Sources/SwiftAgentMacros/SwiftAgentMacroPlugin.swift
@@ -1,0 +1,12 @@
+// By Dennis Müller
+
+import SwiftCompilerPlugin
+import SwiftSyntaxMacros
+
+@main
+struct SwiftAgentMacroPlugin: CompilerPlugin {
+  let providingMacros: [Macro.Type] = [
+    ResolvableToolMacro.self,
+    SwiftAgentSessionMacro.self,
+  ]
+}
